Change loop to just make all squares based on area
Original loop consisted of one outer loop that made the rows,
and a nested one that made the columns. This was unneccessary,
flex-wrap would automatically wrap the divs onto a new line when
they reached the end of the gridBox container.

Also updated CSS to select divs inside gridBox and not "square".
  - That was a typo meant to be ".square" but even that was
    an unneccessary class to have as nothing but the divs should
    be located in the gridBox container (as of now).