Fix bug: no hover event if grid is resized
The event listener for mousover is only applied once: when the
document loads. Therfore, when the grid is resized using the
button, the new div squares don't inherit the event listener. To
fix this, I just put the event listener in a function and called
it twice: at the beginning and when the grid size button is clicked.